+++ ace/__init__.py
@@ -1,7 +1,20 @@
 # vim: sw=4:ts=4:et
 
+import functools
+
 import pytz
 import tzlocal
 
 # local timezone
 LOCAL_TIMEZONE = pytz.timezone(tzlocal.get_localzone().zone)
+
+
+def coreapi(func):
+    """Specifies the given function is a core api function."""
+
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        return func(*args, **kwargs)
+
+    wrapper.__coreapi__ = True
+    return wrapper
