+++ ace/logging.py
@@ -1,0 +1,7 @@
+# vim: sw=4:ts=4:et:cc=120
+
+import logging
+
+
+def get_logger():
+    return logging.getLogger("ace")
