+++ ace/system/base/tokenization.py
@@ -1,0 +1,8 @@
+# vim: ts=4:sw=4:et:cc=120
+#
+#
+#
+
+
+class TokenizationBaseInterface:
+    pass
