+++ ace/system/database/analysis_request.py
@@ -9,7 +9,7 @@ from typing import Optional, Union
 import ace
 
 from ace.analysis import Observable, AnalysisModuleType
-from ace.system import ACESystem
+from ace.system.base import AnalysisRequestTrackingBaseInterface
 from ace.system.database.schema import AnalysisRequestTracking, analysis_request_links
 from ace.constants import TRACKING_STATUS_ANALYZING, EVENT_AR_EXPIRED
 from ace.system.analysis_request import AnalysisRequest
@@ -19,7 +19,7 @@ from ace.exceptions import UnknownAnalysisModuleTypeError
 from sqlalchemy import and_, text
 
 
-class DatabaseAnalysisRequestTrackingInterface(ACESystem):
+class DatabaseAnalysisRequestTrackingInterface(AnalysisRequestTrackingBaseInterface):
     # if we switched to TRACKING_STATUS_ANALYZING then we start the expiration timer
     async def i_track_analysis_request(self, request: AnalysisRequest):
         # XXX we're using server-side time instead of database time
