+++ ace/system/redis/events.py
@@ -5,11 +5,12 @@ import threading
 from typing import Optional
 
 from ace.data_model import Event, custom_json_encoder
-from ace.system import get_logger, ACESystem
+from ace.logging import get_logger
+from ace.system.base import EventBaseInterface
 from ace.system.events import EventHandler
 
 
-class RedisEventInterface(ACESystem):
+class RedisEventInterface(EventBaseInterface):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._rc = None
