+++ ace/system/remote/analysis_module.py
@@ -3,10 +3,10 @@
 from typing import Union
 
 from ace.analysis import AnalysisModuleType
-from ace.system import ACESystem
+from ace.system.base import AnalysisModuleTrackingBaseInterface
 
 
-class RemoteAnalysisModuleTrackingInterface(ACESystem):
+class RemoteAnalysisModuleTrackingInterface(AnalysisModuleTrackingBaseInterface):
     async def register_analysis_module_type(self, amt: AnalysisModuleType) -> AnalysisModuleType:
         return await self.get_api().register_analysis_module_type(amt)
