+++ ace/system/remote/analysis_request.py
@@ -3,10 +3,10 @@
 from typing import Union
 
 from ace.analysis import AnalysisModuleType
-from ace.system import ACESystem
-from ace.system.analysis_request import AnalysisRequest
+from ace.system.base import AnalysisRequestTrackingBaseInterface
+from ace.system.base.analysis_request import AnalysisRequest
 
 
-class RemoteAnalysisRequestTrackingInterface(ACESystem):
+class RemoteAnalysisRequestTrackingInterface(AnalysisRequestTrackingBaseInterface):
     async def process_analysis_request(self, ar: AnalysisRequest):
         return await self.get_api().process_analysis_request(ar)
