+++ ace/system/remote/work_queue.py
@@ -4,10 +4,10 @@ from typing import Union, Optional
 
 from ace.analysis import AnalysisModuleType
 from ace.system.analysis_request import AnalysisRequest
-from ace.system import ACESystem
+from ace.system.base import WorkQueueBaseInterface
 
 
-class RemoteWorkQueueManagerInterface(ACESystem):
+class RemoteWorkQueueManagerInterface(WorkQueueBaseInterface):
     async def get_next_analysis_request(
         self,
         owner_uuid: str,
